Add base MipAPI base class
